+++ packages/iocuak/src/binding/models/api/BaseBindingApi.ts
@@ -1,0 +1,7 @@
+import { ServiceId } from '../../../common/models/domain/ServiceId';
+import { BindingApiType } from './BindingApiType';
+
+export interface BaseBindingApi {
+  id: ServiceId;
+  bindingType: BindingApiType;
+}