+++ packages/iocuak/src/binding/models/api/BindingApiType.ts
@@ -1,0 +1,4 @@
+export enum BindingApiType {
+  type,
+  value,
+}