+++ packages/iocuak/src/binding/models/api/ValueBindingApi.ts
@@ -1,0 +1,7 @@
+import { BaseBindingApi } from './BaseBindingApi';
+import { BindingApiType } from './BindingApiType';
+
+export interface ValueBindingApi<TInstance = unknown> extends BaseBindingApi {
+  bindingType: BindingApiType.value;
+  value: TInstance;
+}