+++ packages/iocuak/src/metadata/models/api/InjectableOptionsApi.ts
@@ -1,0 +1,7 @@
+import { ServiceId } from '../../../common/models/domain/ServiceId';
+import { TaskScopeApi } from '../../../task/models/api/TaskScopeApi';
+
+export interface InjectableOptionsApi {
+  id?: ServiceId;
+  scope?: TaskScopeApi;
+}